wip: break Query into SqlQuery and PhpQuery
